Refactor File Uploads
Files uploading and submitting.

refs: MBL-12078
affects: student
release note: none

+++ Core/Core/Extensions/Int64Extensions.swift
@@ -16,7 +16,7 @@
 
 import Foundation
 
-extension Int64 {
+extension Int {
     static let humanReadableFormatter: ByteCountFormatter = {
         let f = ByteCountFormatter()
         f.countStyle = .file
@@ -24,6 +24,6 @@ extension Int64 {
     }()
 
     public var humanReadableFileSize: String {
-        return Int64.humanReadableFormatter.string(fromByteCount: self)
+        return Int.humanReadableFormatter.string(fromByteCount: Int64(self))
     }
 }
